requirements.txt
streamlit
pyMSAviz==0.5.0
biopython>=1.79
matplotlib>=3.3.4
